# ENH: Optimization of spatial join