This PR increases the performance of geopandas' spatial join function (sjoin.py). The work is based largely on @perrygeo's original work, but introduces a number of optimizations that (a) improve the speed of the join, and (b) reduce memory consumption, allowing the function to handle larger datasets.

Performance increases are primarily achieved by vectorizing operations that were previously handled through iteration (e.g. querying the Rtree index, checking binary predicates). Performance increases are also obtained by generating 'prepared geometry' instances using the `shapely.prepared` package.

Sample speed tests are shown below. The original sjoin function is indicated by `tools.sjoin` while the new sjoin function is indicated by `fast_sjoin`. Join speed is increased by about an order of magnitude for a small dataset (23,000 polygons), by about 14x for a medium dataset (200,000 polygons) and by almost 20x for a large dataset (600,000 polygons).

#### Test 1: ~23000x1500 polygons

```
In [2]: (l_df.shape, r_df.shape)
Out[2]: ((1526, 10), (22669, 9))

In [3]: %timeit fast_sjoin(l_df, r_df, op='contains')
1 loops, best of 3: 4.13 s per loop

In [4]: %timeit tools.sjoin(l_df, r_df, op='contains')
1 loops, best of 3: 53.9 s per loop

```

#### Test 2: ~200000x1500 polygons

```
In [2]: (l_df.shape, r_df.shape)
Out[2]: ((1526, 10), (201853, 9))

In [3]: %timeit fast_sjoin(l_df, r_df, op='contains')
1 loops, best of 3: 31.8 s per loop

In [4]: %timeit tools.sjoin(l_df, r_df, op='contains')
1 loops, best of 3: 7min 27s per loop

```

#### Test 3: ~600000x1500 polygons

```
In [2]: (l_df.shape, r_df.shape)
Out[2]: ((1526, 10), (598502, 9))

In [3]: %timeit fast_sjoin(l_df, r_df, op='contains')
1 loops, best of 3: 2min 9s per loop

In [4]: %timeit tools.sjoin(l_df, r_df, op='contains')
1 loops, best of 3: 37min 12s per loop
```

Let me know what you think!

Thanks,
MDB

Note: special cases (namely, point-in-polygon joins) can probably be optimized further using the `shapely.vectorized` module. However, I haven't implemented these special cases yet.

Also note: I am in the process of tweaking the spatial join tests because the output format is currently a little bit different than in the previous implementation.